@@ -23,6 +23,7 @@
   },
   "repository": "github:graphile/postgraphile",
   "main": "build/index.js",
+  "types": "build/index.d.ts",
   "bin": {
     "postgraphile": "build/postgraphile/cli.js"
   },
@@ -34,6 +35,10 @@
     "prepack": "./scripts/build"
   },
   "dependencies": {
+    "@types/graphql": "^0.8.2",
+    "@types/jsonwebtoken": "<7.2.1",
+    "@types/koa": "2.0.44",
+    "@types/pg": "^7.4.10",
     "body-parser": "^1.15.2",
     "chalk": "^1.1.3",
     "commander": "^2.9.0",
@@ -54,19 +59,23 @@
   },
   "devDependencies": {
     "@types/chalk": "^0.4.30",
-    "@types/change-case": "2.3.1",
     "@types/commander": ">=2.3.30 <2.12.0",
     "@types/debug": "0.0.29",
-    "@types/graphql": "^0.8.2",
+    "@types/finalhandler": "0.0.32",
+    "@types/http-errors": "1.6.1",
     "@types/jest": "^16.0.4",
-    "@types/jsonwebtoken": "<7.2.1",
-    "@types/lodash": ">=4.14.30 <4.14.57",
+    "@types/lodash": "^4.14.110",
+    "@types/lru-cache": "4.1.1",
     "@types/node": "^7.0.4",
+    "@types/parseurl": "1.3.1",
+    "@types/send": "0.14.4",
     "codemirror-graphql": "^0.6.12",
     "connect": "^3.5.0",
     "express": "^4.14.0",
     "graphiql": "^0.11.2",
     "jest": "^18.1.0",
+    "koa": "^2.5.1",
+    "koa-compress": "^3.0.0",
     "nodemon": "^1.11.0",
     "pg-minify": "~0.5.3",
     "react": "^15.3.2",
@@ -75,8 +84,8 @@
     "source-map-support": "^0.4.6",
     "supertest": "^2.0.1",
     "ts-node": "^2.0.0",
-    "tslint": "^4.2.0",
-    "typescript": "2.1.5"
+    "tslint": "^5.10.0",
+    "typescript": "^2.9.2"
   },
   "jest": {
     "transform": {
