@@ -1,47 +1,50 @@
-/* eslint-disable */// Because we use tslint
+/* tslint:disable:no-any */
 import {
   join as joinPath,
   resolve as resolvePath,
   relative as relativePath,
 } from 'path'
 import { readFile } from 'fs'
-import { IncomingMessage, ServerResponse } from 'http'
 import {
   Source,
   parse as parseGraphql,
   validate as validateGraphql,
   execute as executeGraphql,
-  getOperationAST,
   formatError as defaultFormatError,
+  GraphQLError,
+  GraphQLSchema,
   print as printGraphql,
   specifiedRules,
+  DocumentNode,
 } from 'graphql'
 import { extendedFormatError } from '../extendedFormatError'
-import { $$pgClient } from '../../postgres/inventory/pgClientFromContext'
-import renderGraphiQL from './renderGraphiQL'
-import debugPgClient from './debugPgClient'
-import setupServerSentEvents from './setupServerSentEvents'
-import setupPgClientTransaction from '../setupPgClientTransaction'
-import withPostGraphileContext from '../withPostGraphileContext'
-import mapKeys from 'lodash/mapKeys'
+import { IncomingMessage, ServerResponse } from 'http'
+import { isKoaApp, middleware as koaMiddleware } from './koaMiddleware'
 import { pluginHookFromOptions } from '../pluginHook'
+import { PostGraphile } from '../../interfaces'
+import HttpRequestHandler = PostGraphile.HttpRequestHandler
+import ICreateRequestHandler = PostGraphile.ICreateRequestHandler
+import mixed = PostGraphile.mixed
+import setupServerSentEvents from './setupServerSentEvents'
+import withPostGraphileContext from '../withPostGraphileContext'
+import { Context as KoaContext } from 'koa'
 
-const chalk = require('chalk')
-const Debugger = require('debug') // tslint:disable-line variable-name
-const httpError = require('http-errors')
-const parseUrl = require('parseurl')
-const finalHandler = require('finalhandler')
-const bodyParser = require('body-parser')
-const sendFile = require('send')
-const LRU = require('lru-cache')
-const crypto = require('crypto')
+import chalk = require('chalk')
+import Debugger = require('debug') // tslint:disable-line variable-name
+import httpError = require('http-errors')
+import parseUrl = require('parseurl')
+import finalHandler = require('finalhandler')
+import bodyParser = require('body-parser')
+import sendFile = require('send')
+import LRU = require('lru-cache')
+import crypto = require('crypto')
 
-const calculateQueryHash = queryString => crypto.createHash('sha1').update(queryString).digest('base64')
+const calculateQueryHash = (queryString: string): string => crypto.createHash('sha1').update(queryString).digest('base64')
 
 // Fast way of checking if an object is empty,
 // faster than `Object.keys(value).length === 0`
 const hasOwnProperty = Object.prototype.hasOwnProperty
-function isEmpty(value) {
+function isEmpty(value: any): boolean {
   for (const key in value) {
     if (hasOwnProperty.call(value, key)) {
       return false
@@ -52,8 +55,8 @@ function isEmpty(value) {
 
 const { POSTGRAPHILE_ENV } = process.env
 
-const debugGraphql = new Debugger('postgraphile:graphql')
-const debugRequest = new Debugger('postgraphile:request')
+const debugGraphql = Debugger('postgraphile:graphql')
+const debugRequest = Debugger('postgraphile:request')
 
 export const graphiqlDirectory = resolvePath(__dirname, '../graphiql/public')
 
@@ -76,10 +79,8 @@ const favicon = new Promise((resolve, reject) => {
 /**
  * The GraphiQL HTML file as a string. We need it to be a string, because we
  * will use a regular expression to replace some variables.
- *
- * @type {Promise<string>}
  */
-const origGraphiqlHtml = new Promise((resolve, reject) => {
+const origGraphiqlHtml: Promise<string> = new Promise((resolve, reject) => {
   readFile(
     resolvePath(__dirname, '../graphiql/public/index.html'),
     'utf8',
@@ -94,15 +95,15 @@ const origGraphiqlHtml = new Promise((resolve, reject) => {
  * We need to be able to share the withPostGraphileContext logic between HTTP
  * and websockets
  */
-const withPostGraphileContextFromReqResGenerator = options => {
+function withPostGraphileContextFromReqResGenerator(options: ICreateRequestHandler):
+  (req: IncomingMessage, res: ServerResponse, moreOptions: any, fn: (ctx: mixed) => any) => Promise<any> {
   const {
-    getGqlSchema,
     pgSettings,
     jwtSecret,
     additionalGraphQLContextFromRequest,
   } = options
   return async (req, res, moreOptions, fn) => {
-    const jwtToken = options.jwtSecret ? getJwtToken(req) : null
+    const jwtToken = jwtSecret ? getJwtToken(req) : null
     const additionalContext =
       typeof additionalGraphQLContextFromRequest === 'function'
         ? await additionalGraphQLContextFromRequest(req, res)
@@ -126,19 +127,21 @@ const withPostGraphileContextFromReqResGenerator = options => {
 }
 
 /**
- * Creates a GraphQL request handler, this is untyped besides some JSDoc types
- * for intellisense.
+ * Creates a GraphQL request handler that can support many different `http` frameworks, including:
  *
- * @param {GraphQLSchema} graphqlSchema
+ * - Native Node.js `http`.
+ * - `connect`.
+ * - `express`.
+ * - `koa` (2.0).
  */
-export default function createPostGraphileHttpRequestHandler(options) {
+export default function createPostGraphileHttpRequestHandler(options: ICreateRequestHandler): HttpRequestHandler {
   const MEGABYTE = 1024 * 1024
   const {
     getGqlSchema,
     pgPool,
     pgSettings,
     pgDefaultRole,
-    queryCacheMaxSize = 100 * MEGABYTE,
+    queryCacheMaxSize = 50 * MEGABYTE,
   } = options
   const pluginHook = pluginHookFromOptions(options)
 
@@ -173,7 +176,7 @@ export default function createPostGraphileHttpRequestHandler(options) {
 
   // Formats an error using the default GraphQL `formatError` function, and
   // custom formatting using some other options.
-  const formatError = error => {
+  const formatError = (error: GraphQLError) => {
     // Get the appropriate formatted error object, including any extended error
     // fields if the user wants them.
     const formattedError =
@@ -185,17 +188,17 @@ export default function createPostGraphileHttpRequestHandler(options) {
     // formatted error.
     if (options.showErrorStack)
       formattedError.stack =
-        options.showErrorStack === 'json'
+        error.stack != null && options.showErrorStack === 'json'
           ? error.stack.split('\n')
           : error.stack
 
     return formattedError
   }
 
-  const DEFAULT_HANDLE_ERRORS = errors => errors.map(formatError)
+  const DEFAULT_HANDLE_ERRORS = (errors: Array<GraphQLError>) => errors.map(formatError)
   const handleErrors = options.handleErrors || DEFAULT_HANDLE_ERRORS
 
-  function convertKoaBodyParserToConnect(req, res, next) {
+  function convertKoaBodyParserToConnect(req: any, _res: any, next: any): any {
     if (req._koaCtx && req._koaCtx.request && req._koaCtx.request.body) {
       req._body = true
       req.body = req._koaCtx.request.body
@@ -221,7 +224,10 @@ export default function createPostGraphileHttpRequestHandler(options) {
 
   // We'll turn this into one function now so it can be better JIT optimised
   const bodyParserMiddlewaresComposed = bodyParserMiddlewares.reduce(
-    (parent, fn) => {
+    (
+      parent: (req: IncomingMessage, res: ServerResponse, next: (err?: Error) => void) => void,
+      fn: (req: IncomingMessage, res: ServerResponse, next: (err?: Error) => void) => void,
+    ): (req: IncomingMessage, res: ServerResponse, next: (err?: Error) => void) => void => {
       return (req, res, next) => {
         parent(req, res, error => {
           if (error) {
@@ -231,12 +237,12 @@ export default function createPostGraphileHttpRequestHandler(options) {
         })
       }
     },
-    (req, res, next) => next(),
+    (_req: IncomingMessage, _res: ServerResponse, next: (err?: Error) => void) => next(),
   )
 
   // And we really want that function to be await-able
-  const parseBody = (req, res) => new Promise((resolve, reject) => {
-    bodyParserMiddlewaresComposed(req, res, error => {
+  const parseBody = (req: IncomingMessage, res: ServerResponse) => new Promise((resolve, reject) => {
+    bodyParserMiddlewaresComposed(req, res, (error: Error) => {
       if (error) {
         reject(error)
       } else {
@@ -264,15 +270,24 @@ export default function createPostGraphileHttpRequestHandler(options) {
 
   // Typically clients use static queries, so we can cache the parse and
   // validate stages for when we see the same query again. Limit the store size
-  // to 100MB (or queryCacheMaxSize) so it doesn't consume too much RAM.
+  // to 50MB-worth of queries (or queryCacheMaxSize) so it doesn't consume too
+  // much RAM.
   const SHA1_BASE64_LENGTH = 28
+  type CacheEntry = {
+    queryDocumentAst: DocumentNode,
+    validationErrors: Array<GraphQLError>,
+    length: number,
+  }
   const queryCache = LRU({
     max: queryCacheMaxSize,
-    length: (n, key) => n.length + SHA1_BASE64_LENGTH,
+    length: (n: CacheEntry, _key: string) => n.length + SHA1_BASE64_LENGTH,
   })
 
-  let lastGqlSchema
-  const parseQuery = (gqlSchema, queryString) => {
+  let lastGqlSchema: GraphQLSchema
+  const parseQuery = (gqlSchema: GraphQLSchema, queryString: string): {
+    queryDocumentAst: DocumentNode,
+    validationErrors: Array<GraphQLError>,
+  } => {
     if (gqlSchema !== lastGqlSchema) {
       queryCache.reset()
       lastGqlSchema = gqlSchema
@@ -282,8 +297,8 @@ export default function createPostGraphileHttpRequestHandler(options) {
     // attempting to exhaust our memory.
     const canCache = queryString.length < 100000
 
-    const hash = canCache && calculateQueryHash(queryString)
-    const result = canCache && queryCache.get(hash)
+    const hash = canCache ? calculateQueryHash(queryString) : null
+    const result = canCache ? queryCache.get(hash!) : null
     if (result) {
       return result
     } else {
@@ -303,9 +318,9 @@ export default function createPostGraphileHttpRequestHandler(options) {
 
       // Validate our GraphQL query using given rules.
       const validationErrors = validateGraphql(gqlSchema, queryDocumentAst, staticValidationRules)
-      const cacheResult = { queryDocumentAst, validationErrors }
+      const cacheResult: CacheEntry = { queryDocumentAst, validationErrors, length: queryString.length }
       if (canCache) {
-        queryCache.set(hash, cacheResult)
+        queryCache.set(hash!, cacheResult)
       }
       return cacheResult
     }
@@ -315,11 +330,8 @@ export default function createPostGraphileHttpRequestHandler(options) {
    * The actual request handler. It’s an async function so it will return a
    * promise when complete. If the function doesn’t handle anything, it calls
    * `next` to let the next middleware try and handle it.
-   *
-   * @param {IncomingMessage} req
-   * @param {ServerResponse} res
    */
-  const requestHandler = async (incomingReq, res, next) => {
+  const requestHandler = async (incomingReq: IncomingMessage, res: ServerResponse, next: (err?: Error) => void) => {
     // You can use this hook either to modify the incoming request or to tell
     // PostGraphile not to handle the request further (return null). NOTE: if
     // you return `null` from this hook then you are also responsible for
@@ -337,7 +349,7 @@ export default function createPostGraphileHttpRequestHandler(options) {
     if (options.enableCors || POSTGRAPHILE_ENV === 'development')
       addCORSHeaders(res)
 
-    const { pathname } = parseUrl(req)
+    const { pathname = '' } = parseUrl(req) || {}
     const isGraphqlRoute = pathname === graphqlRoute
 
     // ========================================================================
@@ -426,11 +438,12 @@ export default function createPostGraphileHttpRequestHandler(options) {
 
         // Sends the asset at this path. Defaults to a `statusCode` of 200.
         res.statusCode = 200
-        if (req._koaCtx) {
-          req._koaCtx.compress = false
+        const koaCtx = (req as object)['_koaCtx']
+        if (koaCtx) {
+          koaCtx.compress = false
         }
         await new Promise((resolve, reject) => {
-          const stream = sendFile(req, assetPathRelative, {
+          sendFile(req, assetPathRelative, {
             index: false,
             root: graphiqlDirectory,
             dotfiles: 'ignore',
@@ -519,10 +532,10 @@ export default function createPostGraphileHttpRequestHandler(options) {
     // The `result` will be used at the very end in our `finally` block.
     // Statements inside the `try` will assign to `result` when they get
     // a result. We also keep track of `params`.
-    let paramsList
-    let results
+    let paramsList: any
+    let results: Array<{data?: any, errors?: Array<GraphQLError>, statusCode?: number}> = []
     const queryTimeStart = !options.disableQueryLog && process.hrtime()
-    let pgRole
+    let pgRole: string
 
     if (debugRequest.enabled) debugRequest('GraphQL query request has begun.')
     let returnArray = false
@@ -556,7 +569,8 @@ export default function createPostGraphileHttpRequestHandler(options) {
       // - `variables`: An optional JSON object containing GraphQL variables.
       // - `operationName`: The optional name of the GraphQL operation we will
       //   be executing.
-      paramsList = typeof req.body === 'string' ? { query: req.body } : req.body
+      const body: string | object = (req as any).body
+      paramsList = typeof body === 'string' ? { query: body } : body
 
       // Validate our paramsList object a bit.
       if (paramsList == null)
@@ -582,9 +596,9 @@ export default function createPostGraphileHttpRequestHandler(options) {
         paramsList = [paramsList]
       }
       paramsList = pluginHook('postgraphile:httpParamsList', paramsList, { options, req, res, returnArray, httpError })
-      results = await Promise.all(paramsList.map(async (params) => {
-        let queryDocumentAst
-        let result
+      results = await Promise.all(paramsList.map(async (params: any) => {
+        let queryDocumentAst: DocumentNode
+        let result: any
         let meta = {}
         try {
           if (!params.query) throw httpError(400, 'Must provide a query string.')
@@ -624,7 +638,7 @@ export default function createPostGraphileHttpRequestHandler(options) {
               `Operation name must be a string, not '${typeof params.operationName}'.`,
             )
 
-          let validationErrors
+          let validationErrors: Array<GraphQLError>
           ({ queryDocumentAst, validationErrors } = parseQuery(gqlSchema, params.query))
 
           if (validationErrors.length === 0) {
@@ -664,7 +678,7 @@ export default function createPostGraphileHttpRequestHandler(options) {
               queryDocumentAst,
               variables: params.variables,
               operationName: params.operationName,
-            }, graphqlContext => {
+            }, (graphqlContext: any) => {
               pgRole = graphqlContext.pgRole
               return executeGraphql(
                 gqlSchema,
@@ -687,19 +701,22 @@ export default function createPostGraphileHttpRequestHandler(options) {
         } finally {
           // Format our errors so the client doesn’t get the full thing.
           if (result && result.errors) {
-            result.errors = handleErrors(result.errors, req, res)
+            result.errors = (handleErrors as any)(result.errors, req, res)
           }
           if (!isEmpty(meta)) {
             result.meta = meta
           }
           // Log the query. If this debugger isn’t enabled, don’t run it.
-          if (!options.disableQueryLog && queryDocumentAst) {
+          if (
+            !options.disableQueryLog &&
+            queryDocumentAst! /* `!` is not strictly true, but stops TS complaining. */
+          ) {
             setTimeout(() => {
               const prettyQuery = printGraphql(queryDocumentAst)
                 .replace(/\s+/g, ' ')
                 .trim()
               const errorCount = (result.errors || []).length
-              const timeDiff = process.hrtime(queryTimeStart)
+              const timeDiff = queryTimeStart && process.hrtime(queryTimeStart)
               const ms =
                 Math.round((timeDiff[0] * 1e9 + timeDiff[1]) * 10e-7 * 100) / 100
 
@@ -735,14 +752,14 @@ export default function createPostGraphileHttpRequestHandler(options) {
       // Finally, we send the client the results.
       if (!returnArray) {
         if (res.statusCode === 200 && results[0].statusCode) {
-          res.statusCode = results[0].statusCode
+          res.statusCode = results[0].statusCode!
         }
         delete results[0].statusCode
       }
 
       res.setHeader('Content-Type', 'application/json; charset=utf-8')
 
-      res.end(JSON.stringify(returnArray ? results : results[0]))
+      res.end(JSON.stringify(returnArray ? results : results[0]!))
 
       if (debugRequest.enabled) debugRequest('GraphQL ' + (returnArray ? 'queries' : 'query') + ' request finished.')
 
@@ -760,43 +777,20 @@ export default function createPostGraphileHttpRequestHandler(options) {
    * - `express`.
    * - `koa` (2.0).
    */
-  const middleware = (a, b, c) => {
+  const middleware: any = (a: any, b: any, c: any) => {
     // If are arguments look like the arguments to koa middleware, this is
     // `koa` middleware.
-    if (a.req && a.res && typeof b === 'function') {
+    if (isKoaApp(a, b)) {
       // Set the correct `koa` variable names…
-      const ctx = a
-      const next = b
-
-      // Hack the req object so we can get back to ctx
-      ctx.req._koaCtx = ctx
-
-      const oldEnd = ctx.res.end
-      ctx.res.end = (body) => {
-        ctx.response.body = body
-      }
-
-      // Execute our request handler. If an error is thrown, we don’t call
-      // `next` with an error. Instead we return the promise and let `koa`
-      // handle the error.
-      return (async () => {
-        let result
-        try {
-          result = await requestHandler(ctx.req, ctx.res, next)
-        } finally {
-          ctx.res.end = oldEnd
-          if (ctx.res.statusCode && ctx.res.statusCode !== 200) {
-            ctx.response.status = ctx.res.statusCode
-          }
-        }
-        return result
-      })()
+      const ctx = (a as KoaContext)
+      const next = (b as (err?: Error) => Promise<any>)
+      return koaMiddleware(ctx, next, requestHandler)
     } else {
       // Set the correct `connect` style variable names. If there was no `next`
       // defined (likely the case if the client is using `http`) we use the
       // final handler.
-      const req = a
-      const res = b
+      const req = (a as IncomingMessage)
+      const res = (b as ServerResponse)
       const next = c || finalHandler(req, res)
 
       // Execute our request handler.
@@ -814,7 +808,7 @@ export default function createPostGraphileHttpRequestHandler(options) {
   middleware.formatError = formatError
   middleware.pgPool = pgPool
   middleware.withPostGraphileContextFromReqRes = withPostGraphileContextFromReqRes
-  return middleware
+  return (middleware as HttpRequestHandler)
 }
 
 /**
@@ -829,7 +823,7 @@ export default function createPostGraphileHttpRequestHandler(options) {
  *
  * [1]: http://www.html5rocks.com/static/images/cors_server_flowchart.png
  */
-function addCORSHeaders(res) {
+function addCORSHeaders(res: ServerResponse): void {
   res.setHeader('Access-Control-Allow-Origin', '*')
   res.setHeader('Access-Control-Request-Method', 'HEAD, GET, POST')
   res.setHeader(
@@ -847,6 +841,13 @@ function addCORSHeaders(res) {
       'Content-Type',
       'Content-Length',
     ].join(', '),
+  )
+}
+
+function createBadAuthorizationHeaderError(): httpError.HttpError {
+  return httpError(
+    400,
+    'Authorization header is not of the correct bearer scheme format.',
   )
 }
 
@@ -876,8 +877,9 @@ const authorizationBearerRex = /^\s*bearer\s+([a-z0-9\-._~+/]+=*)\s*$/i
  * @param {IncomingMessage} request
  * @returns {string | null}
  */
-function getJwtToken(request) {
+function getJwtToken(request: IncomingMessage): string | null {
   const { authorization } = request.headers
+  if (Array.isArray(authorization)) throw createBadAuthorizationHeaderError()
 
   // If there was no authorization header, just return null.
   if (authorization == null) return null
@@ -886,11 +888,7 @@ function getJwtToken(request) {
 
   // If we did not match the authorization header with our expected format,
   // throw a 400 error.
-  if (!match)
-    throw httpError(
-      400,
-      'Authorization header is not of the correct bearer scheme format.',
-    )
+  if (!match) throw createBadAuthorizationHeaderError()
 
   // Return the token from our match.
   return match[1]
