@@ -1,6 +1,7 @@
 // TODO: Refactor this module, it has code smell…
-
-import { Client } from 'pg'
+import { PostGraphile } from '../../interfaces'
+import { ClientBase, PoolClient } from 'pg'
+import mixed = PostGraphile.mixed
 
 export const $$pgClient = 'pgClient'
 
@@ -8,7 +9,7 @@ export const $$pgClient = 'pgClient'
  * Retrieves a Postgres client from a context, throwing an error if such a
  * client does not exist.
  */
-export default function getPgClientFromContext(context: mixed): Client {
+export default function getPgClientFromContext(context: mixed): PoolClient {
   if (context == null || typeof context !== 'object')
     throw new Error('Context must be an object.')
 
@@ -17,8 +18,8 @@ export default function getPgClientFromContext(context: mixed): Client {
   if (client == null)
     throw new Error('Postgres client does not exist on the context.')
 
-  if (!(client instanceof Client))
+  if (!(client instanceof ClientBase))
     throw new Error('Postgres client on context is of the incorrect type.')
 
-  return client
+  return client as PoolClient
 }
